get both main buses working
